requirements.txt
httpx
python-telegram-bot